backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "openai>=2.32.0",
    "pdfminer-six==20231228",
    "pydantic>=2.13.1",
    "python-multipart>=0.0.26",
    "starlette>=1.0.0",
    "uvicorn>=0.44.0",
]
